nixos/mysql: run the unit as mysql user
With systemd v242 the runtime directory (/run/mysqld) would be destroyed
before starting the daemon. Since the unit was declared to run as root
user systemd would recreate the runtime directory with those permissions
inbetween PreExecStart and ExecStart. When mysqld would then fork and
change uids it would consequently fail to bind on
/run/mysqld/mysqld.sock since it didn't have permissions for the folder.